e2e tests for importing of delegateProxy contract, refactoring of e2e tests: dividing test cases to separate files, updateABI menu item for delegateProxy type
